Set up the optimisation request logic
Set up the optimisation request logic. Set up the DTOs needed. Currently, a DTO exists that handles the Optimisation run as a whole, the hourly schedule and the included in it production units. An OptimiserService was set up, that handles all the communication with the Optimiser module. Currently, on optimisation request - on /optimmisation endpoint, the module calls the RequestOptimisation method part of the OptimiserService. The method makes a POST request to the optimiser module, sending OptimisationRequestDto, carrying the information about the requested optimisation, the optimiser returns the ready optimisation. The method takes care of the deserialisiation of the object and returns the DTO to the controller, the DTO, sents it back to the requester.

+++ services/rdm/Rdm.Api/Inrastructure/Persistence/PersistenceModels/OptimisationProductionUnitPersistence.cs
@@ -15,5 +15,5 @@ public class OptimisationProductionUnitPersistence: BaseModel
     [Column("production_unit_type")]
     public string ProductionUnitType { get; set; }
     [Column("capacity")]
-    public int Capacity { get; set; }
+    public double Capacity { get; set; }
 }
